refactor(QueryContext): move df method utils to utils module
